+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `UnicornInstall` and `UnicornUninstall` events with `os.queueEvent`.
 - Add the `instdat.gitlab_instance` field in the GitLab provider.
+- Add the `unicorn.provider.local.nothing` provider. This is useful for
+  making packages that don't need to contain anything; for example,
+  packages which function as "groups" to install certain packages
+  at once, or packages for software that is already installed.
 
 ## v1.1.0